+++ picoserve/src/json.rs
@@ -1,3 +1,6 @@
-/// A JSON encoded value. When serializing, the value might be serialized several times during sending, so the value must be serialized in the same way each time.
-/// When values are deserialized, `UNESCAPE_BUFFER_SIZE` is the size of the temporary buffer used for unescaping strings.
-pub struct Json<T, const UNESCAPE_BUFFER_SIZE: usize = 32>(pub T);
+/// A JSON encoded value.  
+/// When serializing, the value might be serialized several times during sending,  
+/// so the value must be serialized in the same way each time.  
+/// When deserializing, only short strings can be unescaped.  
+/// If you want to handle longed escaped strings, use [`JsonWithUnescapeBufferSize`](crate::extract::JsonWithUnescapeBufferSize).
+pub struct Json<T>(pub T);